splitted setups into 2 versions, and added github workflow to automatically push on tag
