+++ .github/workflows/continuous-deployment.yaml
@@ -1,0 +1,52 @@
+name: Continuous Deployment
+
+on:
+  push:
+    branches: [ $default-branch ]
+    tags:
+    - "v*"
+
+env:
+  GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+
+defaults:
+  run:
+    shell: bash
+ 
+jobs:
+  builder:
+    strategy:
+      matrix:
+        os: [macos-10.15, windows-2019, ubuntu-latest]
+    runs-on: ${{ matrix.os }}
+    steps:
+      - name: Check out repository
+        uses: actions/checkout@v3
+      - name: Set tag
+        run: echo "GIT_TAG_NAME=${GITHUB_REF#refs/*/}" >> $GITHUB_ENV
+      - name: Discover python architecture
+        run: |
+          if [ "$RUNNER_OS" == "Windows" ]; then
+            echo "PYTHON_ARCHITECTURE=x86" >> $GITHUB_ENV
+          else
+            echo "PYTHON_ARCHITECTURE=x64" >> $GITHUB_ENV
+          fi
+      - name: Set up python
+        uses: actions/setup-python@v3
+        with:
+          python-version: "3.9.x"
+          architecture: "${{ env.PYTHON_ARCHITECTURE }}"
+      - run: python -VV
+      - name: Update dependencies
+        run: python -m pip install -U pip wheel setuptools, twine
+      - name: Build universal source Archive and wheel
+        run: python setup.py sdist bdist_wheel
+      - name: Build binary Archive
+        run: python setup_binary.py download_pandoc bdist_wheel
+      - name: Publish a Python distribution to PyPI
+        uses: pypa/gh-action-pypi-publish@release/v1
+        with:
+          skip_existing: true
+          verbose: true
+          user: __token__
+          password: ${{ secrets.PYPI_API_TOKEN }}
