+++ .gitignore
@@ -1,3 +1,4 @@
+|.egg-info
 *.swp
 *.pyc
 dist/
