+++ appveyor.yml
@@ -79,7 +79,7 @@ install:
 
 build_script:
   - "poetry run python --version"
-  - "poetry run python setup.py download_pandoc"
+  - "poetry run python setup_binary.py download_pandoc"
   - "poetry run pypandoc\\files\\pandoc --version"
 
 test_script:
