+++ pypandoc_binary.egg-info/SOURCES.txt
@@ -1,0 +1,19 @@
+LICENSE
+MANIFEST.in
+README.md
+filter_test.md
+pyproject.toml
+setup.py
+setup_binary.py
+tests.py
+examples/services.py
+pypandoc/__init__.py
+pypandoc/handler.py
+pypandoc/pandoc_download.py
+pypandoc/py3compat.py
+pypandoc/files/COPYRIGHT.txt
+pypandoc/files/pandoc.exe
+pypandoc_binary.egg-info/PKG-INFO
+pypandoc_binary.egg-info/SOURCES.txt
+pypandoc_binary.egg-info/dependency_links.txt
+pypandoc_binary.egg-info/top_level.txt
